feat(module: confirm): add `AutoFocusButton` options for the open method in confirm service (#2398)
* fix decimal conversion

* Fix Culture Decimal Separator

* feat(ConfirmService) Allow choose AutoFocusButton

* Update components/modal/confirmDialog/ConfirmService.cs


* Fix after add Nullable to ConfirmAutoFocusButton

* Fix format in code

+++ components/core/CssSizeLength.cs
@@ -87,7 +87,7 @@ namespace AntDesign
                 throw new FormatException();
             }
 
-            _value = int.Parse(value.Substring(0, index), CultureInfo.InvariantCulture);
+            _value = decimal.Parse(value.Substring(0, index), CultureInfo.InvariantCulture);
 
             if (index == value.Length)
             {
